feat: include cardinality for requirement facets and switching between entity-based and more specific applicability
Switching between entity-based and more specific applicability allows to generate an entity-based applicability even though more columns are used in the applicability. This makes it possible to generate both classic and advanced information requirements from a single Excel file. Which option is used is defined in the IDS4ALL sheet.